backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
pydantic[email]
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
openai>=1.40.0
celery[redis]>=5.4.0
redis>=5.0.0
aioboto3>=13.0.0
yt-dlp>=2024.10.7
google-api-python-client>=2.147.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-multipart>=0.0.9
